Proper User-Character Relationship - Delete & Get Single Character

+++ udemy-dotnet6-rpg/Services/CharacterService/CharacterService.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using System.Security.Claims;
+using udemy_dotnet6_rpg.DTOS.Character;
 
 namespace udemy_dotnet6_rpg.Services.CharacterService
 {
@@ -40,11 +41,22 @@ namespace udemy_dotnet6_rpg.Services.CharacterService
 
 			try
 			{
-				Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-				_context.Characters.Remove(character);
-				await _context.SaveChangesAsync();
+				Character character = await _context.Characters
+				   .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
-				response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDTO>(c)).ToList();
+				if (character != null)
+				{
+					_context.Characters.Remove(character);
+					await _context.SaveChangesAsync();
+					response.Data = _context.Characters
+					.Where(c => c.User.Id == GetUserId())
+						.Select(c => _mapper.Map<GetCharacterDTO>(c)).ToList();
+				}
+				else
+				{
+					response.Success = false;
+					response.Message = "Character not found";
+				}
 			}
 			catch (Exception ex)
 			{
@@ -68,7 +80,8 @@ namespace udemy_dotnet6_rpg.Services.CharacterService
 		public async Task<ServiceResponse<GetCharacterDTO>> GetCharacterById(int id)
 		{
 			var serviceResponse = new ServiceResponse<GetCharacterDTO>();
-			var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+			var dbCharacter = await _context.Characters
+				.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 			serviceResponse.Data = _mapper.Map<GetCharacterDTO>(dbCharacter);
 			return serviceResponse;
 		}
